Refactor the body package to a better structure.
