+++ src/main/java/seedu/address/model/entity/PhoneNumber.java
@@ -1,4 +1,4 @@
-package seedu.address.model.body;
+package seedu.address.model.entity;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.AppUtil.checkArgument;
