+++ src/main/java/seedu/address/model/entity/Sex.java
@@ -1,4 +1,4 @@
-package seedu.address.model.body;
+package seedu.address.model.entity;
 
 //@@author ambervoong
 /**
