+++ src/main/java/seedu/address/model/entity/body/Religion.java
@@ -1,4 +1,4 @@
-package seedu.address.model.body;
+package seedu.address.model.entity.body;
 
 //@@author ambervoong
 /**
