+++ src/test/java/seedu/address/model/entity/body/PhoneNumberTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.body;
+package seedu.address.model.entity.body;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -7,6 +7,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
+
+import seedu.address.model.entity.PhoneNumber;
 
 //@@author ambervoong
 class PhoneNumberTest {
