+++ src/test/java/seedu/address/model/entity/body/SexTest.java
@@ -1,9 +1,11 @@
-package seedu.address.model.body;
+package seedu.address.model.entity.body;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
 
 import org.junit.jupiter.api.Test;
+
+import seedu.address.model.entity.Sex;
 
 //@@author ambervoong
 class SexTest {
